+++ .gitignore
@@ -24,3 +24,6 @@ coverage/
 ## Added by Auro CLI migration (github-config-and-gitignore-changes)
 demo/*.md
 demo/*.min.js
+
+## CEM
+custom-elements.json
