+++ apiExamples/close-button.html
@@ -6,7 +6,6 @@
   .example::part(close-button) {
     top: var(--ds-size-400);
     right: var(--ds-size-400);
-    color: var(--ds-color-brand-flamingo-500);
   }
 </style>
