+++ apiExamples/close-button.js
