+++ package.json
@@ -24,13 +24,13 @@
     "lit": "^3.3.1"
   },
   "devDependencies": {
-    "@aurodesignsystem/auro-button": "^12.2.0",
-    "@aurodesignsystem/auro-cli": "^3.0.4",
+    "@aurodesignsystem/auro-button": "^12.3.0",
+    "@aurodesignsystem/auro-cli": "^3.5.0",
     "@aurodesignsystem/auro-config": "^1.3.1",
-    "@aurodesignsystem/auro-icon": "^9.1.0",
+    "@aurodesignsystem/auro-icon": "^9.1.1",
     "@aurodesignsystem/auro-library": "^5.5.4",
-    "@aurodesignsystem/design-tokens": "^8.5.0",
-    "@aurodesignsystem/webcorestylesheets": "^10.1.1",
+    "@aurodesignsystem/design-tokens": "^8.7.0",
+    "@aurodesignsystem/webcorestylesheets": "^10.1.4",
     "husky": "^9.1.7"
   },
   "browserslist": [
@@ -53,6 +53,7 @@
   "scripts": {
     "build": "auro build",
     "build:version": "node ./scripts/version.mjs",
+    "./custom-elements.json": "./custom-elements.json",
     "dev": "auro dev -w -s",
     "dev:open": "auro dev -w -s --open",
     "lint": "biome check --no-errors-on-unmatched && stylelint \"./src/**/*.scss\"",
@@ -60,12 +61,18 @@
     "test": "auro test",
     "test:watch": "auro test --watch",
     "test:coverage": "auro test --coverage-report --open",
-    "prepare": "husky"
+    "prepare": "husky",
+    "cem": "auro docs --cem --api"
   },
   "exports": {
     "./package.json": "./package.json",
     "./readme.md": "./README.md",
-    ".": "./dist/registered.js",
+    "./custom-elements.json": "./custom-elements.json",
+    ".": {
+      "module": "./dist/registered.js",
+      "types": "./dist/index.d.ts",
+      "default": "./dist/registered.js"
+    },
     "./demo/*.md": "./demo/*.md",
     "./demo/*.js": "./demo/*.min.js",
     "./class": {
@@ -81,5 +88,6 @@
     "README.md",
     "LICENSE",
     "NOTICE"
-  ]
+  ],
+  "customElements": "custom-elements.json"
 }
