+++ src/iconVersion.js
@@ -1,1 +1,1 @@
-export default '9.1.0'
+export default '9.1.1'
